Split VideoPlayerView.swift into per-component files
Break the 700-line VideoPlayerView.swift into a Player/Views/Video/ group,
matching the repo's one-type-per-file + subfolder convention:
- VideoArtworkView.swift — the SwiftUI artwork-replacement view
- VideoPlayerSurface.swift — VideoPlayerSurface + VideoSurfaceUIView + PlayerLayerView
- VideoPiPCoordinator.swift — the Picture in Picture coordinator
- VideoFullscreenViewController.swift — the fullscreen controller + presenter

Move the file-private videoFullscreenDidDismiss notification (now shared by
the surface and the fullscreen VC) into Notification+BookPlayer.swift.
Pure reorganization — no behavior change.
