+++ BookPlayer/Player/Views/Video/VideoFullscreenViewController.swift
@@ -1,5 +1,5 @@
 //
-//  VideoPlayerView.swift
+//  VideoFullscreenViewController.swift
 //  BookPlayer
 //
 //  Created by Pedro Iñiguez on 15/7/26.
@@ -7,367 +7,8 @@
 //
 
 import AVFoundation
-import AVKit
 import BookPlayerKit
-import SwiftUI
 import UIKit
-
-private extension Notification.Name {
-  /// Posted when the fullscreen video controller finishes tearing down. The inline
-  /// surface shares the same `AVPlayer`; the fullscreen surface takes over its video
-  /// output, so the inline surface must reclaim rendering once fullscreen is gone.
-  static let videoFullscreenDidDismiss = Notification.Name("com.bookplayer.videoFullscreenDidDismiss")
-}
-
-// MARK: - Picture in Picture
-
-/// Coordinates system Picture in Picture for the player screen's video surface.
-///
-/// PiP auto-starts only when the hosting layer is visible and playing at the moment
-/// the app is backgrounded (`canStartPictureInPictureAutomaticallyFromInline`), which
-/// gives the intended behavior for free: leaving the app from the player screen with
-/// a playing video starts PiP; anywhere else in the app (or with the player dismissed,
-/// which detaches the surface) it doesn't.
-@MainActor
-final class VideoPiPCoordinator: NSObject, AVPictureInPictureControllerDelegate, BPLogger {
-  static let shared = VideoPiPCoordinator()
-
-  private var pipController: AVPictureInPictureController?
-  private var bookEndObserver: NSObjectProtocol?
-
-  var isActive: Bool {
-    pipController?.isPictureInPictureActive ?? false
-  }
-
-  /// Gated on device support and the Picture in Picture setting
-  var isEnabled: Bool {
-    AVPictureInPictureController.isPictureInPictureSupported()
-      && UserDefaults.standard.bool(forKey: Constants.UserDefaults.videoPictureInPictureEnabled)
-  }
-
-  override private init() {
-    super.init()
-
-    /// The end of the book/video dismisses the PiP window
-    bookEndObserver = NotificationCenter.default.addObserver(
-      forName: .bookEnd,
-      object: nil,
-      queue: .main
-    ) { _ in
-      Task { @MainActor in
-        Self.shared.stop()
-      }
-    }
-  }
-
-  func host(playerLayer: AVPlayerLayer) {
-    guard isEnabled else {
-      release(playerLayer: playerLayer)
-      return
-    }
-
-    guard pipController?.playerLayer !== playerLayer else { return }
-
-    guard let controller = AVPictureInPictureController(playerLayer: playerLayer) else {
-      Self.logger.error("PiP: failed to create controller (unsupported device?)")
-      return
-    }
-
-    controller.canStartPictureInPictureAutomaticallyFromInline = true
-    controller.delegate = self
-    pipController = controller
-  }
-
-  func release(playerLayer: AVPlayerLayer) {
-    guard pipController?.playerLayer === playerLayer else { return }
-
-    pipController?.delegate = nil
-    pipController = nil
-  }
-
-  func stop() {
-    guard isActive else { return }
-
-    pipController?.stopPictureInPicture()
-  }
-
-  // MARK: AVPictureInPictureControllerDelegate
-
-  nonisolated func pictureInPictureController(
-    _ pictureInPictureController: AVPictureInPictureController,
-    failedToStartPictureInPictureWithError error: Error
-  ) {
-    Self.logger.error("PiP: failed to start: \(error.localizedDescription)")
-  }
-
-  /// The user tapped "return to app" on the PiP window. The player screen is still
-  /// presented (PiP can only have started from it), so there's nothing to restore —
-  /// just report completion so the system finishes stopping PiP. Playback continues.
-  /// Closing the window instead (✕) stops PiP with no restore, and the system pauses
-  /// playback — the intended behavior: closing the video means "stop", not "keep listening".
-  nonisolated func pictureInPictureController(
-    _ pictureInPictureController: AVPictureInPictureController,
-    restoreUserInterfaceForPictureInPictureStopWithCompletionHandler completionHandler: @escaping (Bool) -> Void
-  ) {
-    completionHandler(true)
-  }
-}
-
-// MARK: - Video surface
-
-/// UIView backed by an `AVPlayerLayer`. Using the backing layer (instead of a
-/// manually laid-out sublayer) lets UIView animations resize the video smoothly
-/// during the fullscreen transition.
-private final class PlayerLayerView: UIView {
-  override class var layerClass: AnyClass { AVPlayerLayer.self }
-
-  // swiftlint:disable:next force_cast
-  var playerLayer: AVPlayerLayer { layer as! AVPlayerLayer }
-}
-
-/// Renders the shared player's video tracks: an aspect-fit layer on top, over an
-/// aspect-fill layer blurred to cover the container's empty space.
-final class VideoSurfaceUIView: UIView {
-  private let showsBlurredBackground: Bool
-  /// Whether this surface's layer backs the system Picture in Picture window
-  /// (only the player screen's inline surface should)
-  private let hostsPictureInPicture: Bool
-  private let backgroundVideoView = PlayerLayerView()
-  private let foregroundVideoView = PlayerLayerView()
-  private let blurView = UIVisualEffectView(effect: UIBlurEffect(style: .regular))
-
-  private var lifecycleObservers = [NSObjectProtocol]()
-  private weak var attachedPlayer: AVPlayer?
-
-  init(showsBlurredBackground: Bool, hostsPictureInPicture: Bool = false) {
-    self.showsBlurredBackground = showsBlurredBackground
-    self.hostsPictureInPicture = hostsPictureInPicture
-    super.init(frame: .zero)
-
-    backgroundColor = .black
-
-    if showsBlurredBackground {
-      backgroundVideoView.playerLayer.videoGravity = .resizeAspectFill
-      addPinnedSubview(backgroundVideoView)
-      addPinnedSubview(blurView)
-    }
-
-    foregroundVideoView.playerLayer.videoGravity = .resizeAspect
-    addPinnedSubview(foregroundVideoView)
-
-    registerLifecycleObservers()
-  }
-
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-
-  deinit {
-    lifecycleObservers.forEach(NotificationCenter.default.removeObserver)
-  }
-
-  private func addPinnedSubview(_ subview: UIView) {
-    subview.frame = bounds
-    subview.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-    addSubview(subview)
-  }
-
-  func attach(_ player: AVPlayer?) {
-    attachedPlayer = player
-    connectLayers(to: player)
-
-    if hostsPictureInPicture {
-      if player != nil {
-        VideoPiPCoordinator.shared.host(playerLayer: foregroundVideoView.playerLayer)
-      } else {
-        VideoPiPCoordinator.shared.release(playerLayer: foregroundVideoView.playerLayer)
-      }
-    }
-  }
-
-  /// Force the layers to re-associate with the attached player. Setting the same
-  /// `AVPlayer` on another `AVPlayerLayer` (the fullscreen surface) stops this layer
-  /// from rendering even though its `player` reference is left untouched — so a plain
-  /// `connectLayers` is a no-op here (identity guard). Detaching and reattaching is
-  /// what reclaims the video output.
-  private func reclaimVideoRendering() {
-    guard let player = attachedPlayer else { return }
-
-    foregroundVideoView.playerLayer.player = nil
-    foregroundVideoView.playerLayer.player = player
-    if showsBlurredBackground {
-      backgroundVideoView.playerLayer.player = nil
-      backgroundVideoView.playerLayer.player = player
-    }
-  }
-
-  private func connectLayers(to player: AVPlayer?) {
-    if foregroundVideoView.playerLayer.player !== player {
-      foregroundVideoView.playerLayer.player = player
-    }
-    if showsBlurredBackground, backgroundVideoView.playerLayer.player !== player {
-      backgroundVideoView.playerLayer.player = player
-    }
-  }
-
-  /// A player rendering video through a layer is paused by the system when the app
-  /// moves to the background. Detaching the layers while backgrounded keeps the audio
-  /// playing — a video always continues as audio in the background, like any other item.
-  private func registerLifecycleObservers() {
-    let center = NotificationCenter.default
-
-    lifecycleObservers.append(
-      center.addObserver(
-        forName: UIApplication.didEnterBackgroundNotification,
-        object: nil,
-        queue: .main
-      ) { [weak self] _ in
-        guard let self else { return }
-
-        /// Leave the layers connected when PiP is taking over this surface —
-        /// detaching here would cancel the automatic PiP start
-        if self.hostsPictureInPicture,
-           VideoPiPCoordinator.shared.isEnabled,
-           self.attachedPlayer?.timeControlStatus != .paused {
-          return
-        }
-
-        self.connectLayers(to: nil)
-      }
-    )
-
-    lifecycleObservers.append(
-      center.addObserver(
-        forName: UIApplication.willEnterForegroundNotification,
-        object: nil,
-        queue: .main
-      ) { [weak self] _ in
-        guard let self else { return }
-
-        self.connectLayers(to: self.attachedPlayer)
-      }
-    )
-
-    /// Returning from fullscreen: the fullscreen surface had taken over this player's
-    /// video output, so reclaim it (a no-op for the torn-down fullscreen surface,
-    /// whose player was already detached).
-    lifecycleObservers.append(
-      center.addObserver(
-        forName: .videoFullscreenDidDismiss,
-        object: nil,
-        queue: .main
-      ) { [weak self] _ in
-        self?.reclaimVideoRendering()
-      }
-    )
-  }
-}
-
-struct VideoPlayerSurface: UIViewRepresentable {
-  let player: AVPlayer
-  var showsBlurredBackground = true
-  var hostsPictureInPicture = false
-
-  func makeUIView(context: Context) -> VideoSurfaceUIView {
-    let view = VideoSurfaceUIView(
-      showsBlurredBackground: showsBlurredBackground,
-      hostsPictureInPicture: hostsPictureInPicture
-    )
-    view.attach(player)
-    return view
-  }
-
-  func updateUIView(_ uiView: VideoSurfaceUIView, context: Context) {
-    /// The player instance can be recreated after a media-services reset
-    uiView.attach(player)
-  }
-
-  static func dismantleUIView(_ uiView: VideoSurfaceUIView, coordinator: ()) {
-    uiView.attach(nil)
-  }
-}
-
-// MARK: - Artwork replacement
-
-/// Takes the artwork's place in the player screen when the current chapter is a video
-struct VideoArtworkView: View {
-  let player: AVPlayer
-  /// Receives the container's frame in global coordinates, so the fullscreen
-  /// transition can animate from the video's exact on-screen position
-  let onFullscreenTap: (CGRect) -> Void
-
-  @State private var containerFrame: CGRect = .zero
-  @State private var controlsVisible = true
-
-  var body: some View {
-    ZStack {
-      // 1. Create the square "box"
-      Color.clear
-        .aspectRatio(1, contentMode: .fit)
-
-      VideoPlayerSurface(player: player, hostsPictureInPicture: true)
-        .contentShape(Rectangle())
-        .onTapGesture {
-          withAnimation(.easeInOut(duration: 0.2)) {
-            controlsVisible.toggle()
-          }
-        }
-
-      VStack {
-        HStack(spacing: 12) {
-          Spacer()
-
-          Button {
-            onFullscreenTap(containerFrame)
-          } label: {
-            Image(systemName: "arrow.up.left.and.arrow.down.right")
-              .resizable()
-              .aspectRatio(contentMode: .fit)
-              .frame(width: 20, height: 20)
-              .foregroundColor(.white)
-              .frame(width: 44, height: 44) // Standard touch target size
-              .contentShape(Rectangle())
-          }
-          .shadow(color: .black.opacity(0.5), radius: 3, x: 0, y: 2)
-          .accessibilityLabel(Text("video_fullscreen_title".localized))
-
-          AirplayPicker()
-            .frame(width: 44, height: 44) // Standard touch target size
-            .padding(.trailing, 5)
-            .shadow(color: .black.opacity(0.5), radius: 3, x: 0, y: 2)
-            .accessibilityElement(children: .combine)
-            .accessibilityLabel(Text("audio_source_title"))
-        }
-        .padding(.top, 5)
-        .frame(maxWidth: .infinity)
-
-        Spacer()
-      }
-      .opacity(controlsVisible ? 1 : 0)
-      .allowsHitTesting(controlsVisible)
-      /// Keep VoiceOver from focusing the fullscreen/AirPlay buttons while they're
-      /// hidden — `opacity(0)` alone leaves them in the accessibility tree
-      .accessibilityHidden(!controlsVisible)
-    }
-    .aspectRatio(1, contentMode: .fit)
-    .cornerRadius(12)
-    .clipped()
-    .background(
-      GeometryReader { geometry in
-        Color.clear
-          .onAppear {
-            containerFrame = geometry.frame(in: .global)
-          }
-          .onChange(of: geometry.frame(in: .global)) { _, newFrame in
-            containerFrame = newFrame
-          }
-      }
-    )
-    .frame(maxWidth: .infinity, maxHeight: .infinity)
-  }
-}
-
-// MARK: - Fullscreen
 
 /// YouTube-style fullscreen: the system rotation to landscape IS the entry animation —
 /// the container expands from the video's inline frame alongside the interface
